requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
python-dotenv==1.1.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-google-genai==2.1.4
langchain-openai==0.3.14
langchain-qdrant==0.2.0
langgraph==0.4.5
langgraph-checkpoint-mongodb==0.1.3
pymongo==4.11.3
qdrant-client==1.14.2
SQLAlchemy==2.0.40
openai==1.76.0
tiktoken==0.9.0
